Move to Maintainability analyzers project
