+++ Philips.CodeAnalysis.DuplicateCodeAnalyzer/Philips.CodeAnalysis.DuplicateCodeAnalyzer.csproj
@@ -28,8 +28,6 @@
 1.1.3:
   Report character number in addition to line number, to more easily identify the start of duplication
   Increase max token count to 200
-1.1.4:
-  Introduce PH2136: Avoid duplicate strings
     </PackageReleaseNotes>
     <Copyright>© 2020-2022 Koninklijke Philips N.V.</Copyright>
     <PackageTags>CSharp MsTest Roslyn CodeAnalysis analyzers cpd duplicate Philips</PackageTags>
