+++ Philips.CodeAnalysis.MaintainabilityAnalyzers/Maintainability/AvoidDuplicateStringsAnalyzer.cs
@@ -9,14 +9,14 @@ using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Microsoft.CodeAnalysis.Diagnostics;
 using Philips.CodeAnalysis.Common;
 
-namespace Philips.CodeAnalysis.DuplicateCodeAnalyzer
+namespace Philips.CodeAnalysis.MaintainabilityAnalyzers.Maintainability
 {
 	[DiagnosticAnalyzer(LanguageNames.CSharp)]
 	public class AvoidDuplicateStringsAnalyzer : SingleDiagnosticAnalyzer
 	{
 		private const string Title = @"Avoid Duplicate Strings";
 		private const string MessageFormat = @"Duplicate string found, first location in file {0} at line {1}. Consider moving '{2}' into a constant.";
-		private const string Description = @"Duplicate code is less maintainable";
+		private const string Description = @"Duplicate strings are less maintainable";
 		private const string Category = Categories.Maintainability;
 
 		public AvoidDuplicateStringsAnalyzer() : base(DiagnosticId.AvoidDuplicateStrings, Title, MessageFormat, Description, Category)
@@ -66,7 +66,7 @@ namespace Philips.CodeAnalysis.DuplicateCodeAnalyzer
 							 .Where(token => token.IsKind(SyntaxKind.StringLiteralToken)))
 				{
 					var literalText = literal.Text.Trim('\\', '\"');
-					if(string.IsNullOrWhiteSpace(literalText))
+					if(string.IsNullOrWhiteSpace(literalText) || literalText.Length <= 2)
 					{
 						continue;
 					}
