+++ Philips.CodeAnalysis.MaintainabilityAnalyzers/Philips.CodeAnalysis.MaintainabilityAnalyzers.csproj
@@ -117,6 +117,7 @@
       * Introduce PH2134: Set properties in any order.
       * PH2090 now also uses an AdditionalFile on top of the existing .editorconfig option to configure allowed methods.
       * Introduce PH2135: Match namespace and Assembly Name. This extends the functionality that was previously claimed by PH2006 but never enforced.
+	  * Introduce PH2136: Avoid duplicate strings.
     </PackageReleaseNotes>
 	  <Copyright>© 2019-2023 Koninklijke Philips N.V.</Copyright>
 	  <PackageTags>CSharp Maintainability Roslyn CodeAnalysis analyzers Philips</PackageTags>
